Extract useMyBars from App.tsx
Move the global users/{uid} subscription (joined-bar list) and the
chunked bar-name fetch into useMyBars(user). Drops the now-unused
getDocs and documentId imports from App.tsx.

53/53 tests pass.
